Contains pure C code which does conversion from perl data to C data.
The input map takes the perl encapsulation of a Ns_Conn struct and
pulls out just the pointer to the struct without the perl infrastructure.
Used when converting parameter arguments.

The output map takes a pointer to a Ns_Conn and builds the perl infrastructure
used to carry it within perl. Used when converting return values, and when
passing the aolserver browser connection object (represented by a struct
Ns_Conn) into perl such that it has the name $Aolserver::Ns_Conn::theConn.