In the process of finding a bug where handle is not returned when var goes out of scope.
Also discovered server dies because an attempt is made to free the
returned Ns_Set from the Select() method. Fix for that is to have
a reference-counted place for it which gets zeroed before the
Ns_DbHandle is itself destroyed (err, returned). The docs say you
must not Ns_SetFree() the return value of Select().