try a simpler structure in which the Ns_DbHandle is given perl infrastructure for the selectRowSet when perl infrastructure for the Ns_DbHandle is created.
In this scheme, the select row infrastructure lasts for the life of the
Ns_DbHandle. To go into select loop mode, one would put a pointer to
a Ns_Set structure at the "C" end of the select row infrastructure;
to come out of this mode, put a NULL pointer in the same place.