Seems to be no change in Aolserver/Ns_Conn/Ns_ConnMaps.c.
In nsperl/nsperl.c, this is an attempt to get rid of the warning
that $Aolserver::Ns_Conn::theConn has only been used once, by trying
to nose around in the GV and set the correct bit instead of using
GV_ADDMULTI in the get_sv call on "Aolserver::Ns_Conn::theConn".

Not sure it's working; not sure why it would not work.