Stop setting perl infrastructure pointers to null in order to protect what they point at.
Instead, introduce ownership variable; if true, the perl infrastructure
"owns" the aolserver data structure and should therefore free it;
if false, only free the perl infrastructure, and NOT the aolserver
data.

Also, add warning in case a perl script which is supposed to return
browser content does not; earlier failures to do so caused unbounded
recursion.